RL/pyproject.toml
[project]
name = "rl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "gymnasium>=1.2.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "pillow>=12.1.0",
    "pygame>=2.6.1",
]
